+++ lib/redis_queued_locks/acquier.rb
@@ -278,6 +278,8 @@ module RedisQueuedLocks::Acquier
     # @api private
     # @since 0.1.0
     def locked?(redis_client, lock_name)
+      lock_key = RedisQueuedLocks::Resource.prepare_lock_key(lock_name)
+      redis_client.call('EXISTS', lock_key) == 1
     end
 
     # @param redis_client [RedisClient]
@@ -287,15 +289,101 @@ module RedisQueuedLocks::Acquier
     # @api private
     # @since 0.1.0
     def queued?(redis_client, lock_name)
+      lock_key_queue = RedisQueuedLocks::Resource.prepare_lock_queue(lock_name)
+      redis_client.call('EXISTS', lock_key_queue) == 1
     end
 
     # @param redis_client [RedisClient]
     # @param lock_name [String]
-    # @return [?]
+    # @return [Hash<Symbol,String|Numeric>,NilClass]
+    #   - `nil` is returned when lock key does not exist or expired;
+    #   - result format: {
+    #     lock_key: "rql:lock:your_lockname", # acquired lock key
+    #     acq_id: "rql:acq:process_id/thread_id", # lock acquier identifier
+    #     ts: 123456789, # <locked at> time stamp (epoch)
+    #     ini_ttl: 123456789, # initial lock key ttl (milliseconds),
+    #     rem_ttl: 123456789, # remaining lock key ttl (milliseconds)
+    #   }
     #
     # @api private
     # @since 0.1.0
     def lock_info(redis_client, lock_name)
+      lock_key = RedisQueuedLocks::Resource.prepare_lock_key(lock_name)
+
+      result = redis_client.multi(watch: [lock_key]) do |transact|
+        transact.call('HGETALL', lock_key)
+        transact.call('PTTL', lock_key)
+      end
+
+      if result == nil
+        # NOTE:
+        #   - nil result means that during transaction invocation the lock is changed (CAS):
+        #     - lock is expired;
+        #     - lock is released;
+        #     - lock is expired + re-obtained;
+        nil
+      else
+        hget_cmd_res = result[0]
+        pttl_cmd_res = result[1]
+
+        if hget_cmd_res == {} || pttl_cmd_res == -2 # NOTE: key does not exist
+          nil
+        else
+          # NOTE: the result of MULTI-command is an array of results of each internal command
+          #   - result[0] (HGETALL) (Hash<String,String>)
+          #   - result[1] (PTTL) (Integer)
+          {
+            lock_key: lock_key,
+            acq_id: hget_cmd_res['acq_id'],
+            ts: Integer(hget_cmd_res['ts']),
+            ini_ttl: Integer(hget_cmd_res['ini_ttl']),
+            rem_ttl: ((pttl_cmd_res == -1) ? Infinity : pttl_cmd_res)
+          }
+        end
+      end
+    end
+
+    # Returns an information about the required lock queue by the lock name. The result
+    # represnts the ordered lock request queue that is ordered by score (Redis sets) and shows
+    # lock acquirers and their position in queue. Async nature with redis communcation can lead
+    # the sitation when the queue becomes empty during the queue data extraction. So sometimes
+    # you can receive the lock queue info with empty queue.
+    #
+    # @param redis_client [RedisClient]
+    # @param lock_name [String]
+    # @return [Hash<Symbol,String|Array<Hash<Symbol,String|Float>>,NilClass]
+    #   - `nil` is returned when lock queue does not exist;
+    #   - result format: {
+    #     lock_queue: "rql:lock_queue:your_lock_name", # lock queue key in redis,
+    #     queue: [
+    #       { acq_id: "rql:acq:process_id/thread_id", score: 123 },
+    #       { acq_id: "rql:acq:process_id/thread_id", score: 456 },
+    #     ] # ordered set (by score) with information about an acquier and their position in queue
+    #   }
+    #
+    # @api private
+    # @since 0.1.0
+    def queue_info(redis_client, lock_name)
+      lock_key_queue = RedisQueuedLocks::Resource.prepare_lock_queue(lock_name)
+
+      result = redis_client.pipelined do |pipeline|
+        pipeline.call('EXISTS', lock_key_queue)
+        pipeline.call('ZRANGE', lock_key_queue, '0', '-1', 'WITHSCORES')
+      end
+
+      exists_cmd_res = result[0]
+      zrange_cmd_res = result[1]
+
+      if exists_cmd_res == 1
+        # NOTE: queue existed during the piepline invocation
+        {
+          lock_queue: lock_key_queue,
+          queue: zrange_cmd_res.map { |val| { acq_id: val[0], score: val[1] } }
+        }
+      else
+        # NOTE: queue did not exist during the pipeline invocation
+        nil
+      end
     end
 
     private